Refactor JavaCheck service
